+++ components/ui/components/ProRestful/ProRestful.stories.tsx
@@ -9,6 +9,7 @@ import Button from '@mui/material/Button';
 import IconButton from '@mui/material/IconButton';
 import EditIcon from '@mui/icons-material/Edit';
 import DeleteIcon from '@mui/icons-material/Delete';
+import Card from "@mui/material/Card";
 
 export default {
   title: 'ProRestful',
@@ -171,4 +172,68 @@ export function Default() {
       {...api}
     />
   );
+}
+
+export function ModalMode() {
+    const ref = React.useRef(null);
+    const columnsHelper = createColumnHelper<User>();
+    const columns = React.useMemo(
+        () => [
+            columnsHelper.accessor('id', {
+                size: 1,
+                cell: (info) => info.getValue(),
+            }),
+            columnsHelper.accessor('username', {
+                cell: (info) => info.getValue(),
+            }),
+            columnsHelper.accessor('password', {
+                cell: (info) => info.getValue(),
+            }),
+            columnsHelper.accessor('createdAt', {
+                cell: (info) => info.getValue(),
+            }),
+            columnsHelper.accessor('updatedAt', {
+                cell: (info) => info.getValue(),
+            }),
+            columnsHelper.accessor('deletedAt', {
+                cell: (info) => info.getValue(),
+            }),
+            {
+                accessorKey: 'actions',
+                cell: (info) => {
+                    return (
+                        <>
+                            <IconButton
+                                onClick={() => {
+                                    ref.current.setFocusedRow(info.row.original);
+                                    ref.current.setIsDrawerOpen(true);
+                                }}
+                            >
+                                <EditIcon />
+                            </IconButton>
+                            <IconButton
+                                onClick={() => {
+                                    ref.current.handleDelete(info.row.original);
+                                }}
+                            >
+                                <DeleteIcon />
+                            </IconButton>
+                        </>
+                    );
+                },
+            },
+        ],
+        [],
+    );
+
+    return (
+        <ProRestful
+            ref={ref}
+            mode={"modal"}
+            columns={columns}
+            createForm={({ handleCreate }) => <Card sx={{ width: 500, margin: '40px auto' }}><CreationForm handleCreate={handleCreate}/></Card>}
+            editForm={({ handleEdit, data }) => <Card sx={{ width: 500, margin: '40px auto' }}><EditionForm handleEdit={handleEdit} data={data}/></Card>}
+            {...api}
+        />
+    );
 }